requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
caer==2.0.8
charset-normalizer==3.3.2
click==8.1.7
discord==2.3.2
discord.py==2.3.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mypy==1.5.1
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
numpy==1.25.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
protobuf==4.21.12
pyodbc==5.0.1
typing_extensions==4.7.1
Werkzeug==3.0.1
yarl==1.9.2

